requirements.txt
Pillow
openpyxl
pandas
